+++ .editorconfig
@@ -1,0 +1,22 @@
+# editorconfig.org
+root = true
+
+[*]
+indent_style = space
+indent_size = 4
+end_of_line = lf
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
+
+[*.md]
+trim_trailing_whitespace = false
+
+[*.yml]
+indent_size = 2
+
+[*.java]
+# Intellij IDEA settings
+ij_java_class_count_to_use_import_on_demand = 9999
+ij_java_imports_layout = |,java.**,|,javax.**,|,com.**,|,org.bukkit.**,|,io.**,|,me.**,|,net.**,|,dev.**,|,*,|,$*
+ij_java_names_count_to_use_import_on_demand = 9999
