+++ .github/ISSUE_TEMPLATE/bug-report-cn.yml
@@ -26,7 +26,7 @@ body:
           required: true
         - label: 你没有安装`CS-CoreLib`
           required: true
-        - label: 你已经寻找过[已知问题列表](https://github.com/ybw0014/SlimeTinker-CN/issues)，且没有找到相同的问题
+        - label: 你已经寻找过[已知问题列表](https://github.com/ybw0014/GuizhanCraft/issues)，且没有找到相同的问题
           required: true
 
   - type: textarea
