--- crowdin.yml
@@ -1,3 +1,0 @@
-files:
-  - source: /src/main/resources/lang/zh-CN.yml
-    translation: /src/main/resources/lang/%osx_locale%.yml
