+++ documents/intro-en-US.md
@@ -1,12 +1,9 @@
 # GuizhanCraft
 
-GuizhanCraft is a Slimefun4 Addon made by ybw0014.
+GuizhanCraft is a Slimefun4 addon made by ybw0014.
 
 ## Content
 
 ### Machines
 
-- Backpack Upgrader: Upgrade Backpacks using this machine. (Made for a special feature in a testing branch of Slimefun Chinese version)
-
-### Magic
-
+- Electric Spawner Assembler：You can assemble electric spawner with electric spawner framework and a reinforced spawner.
