+++ documents/intro-zh-CN.md
@@ -6,6 +6,4 @@
 
 ### 机器
 
-- 背包升级器：可以将通过该机器升级背包（为了汉化版测试分支的某个特性而制作的）
-
-### 魔法
+- 电力刷怪笼组装机：可以将电力刷怪笼框架与已修复的刷怪笼组装成电力刷怪笼。
