--- src/main/java/net/guizhanss/guizhancraft/slimefun/GuizhanCraftRecipeTypes.java
@@ -1,14 +1,0 @@
-package net.guizhanss.guizhancraft.slimefun;
-
-import io.github.thebusybiscuit.slimefun4.api.recipes.RecipeType;
-import lombok.experimental.UtilityClass;
-
-/**
- * This class holds all {@link RecipeType} in GuizhanCraft
- *
- * @author ybw0014
- */
-@UtilityClass
-public final class GuizhanCraftRecipeTypes {
-
-}
