--- src/main/java/net/guizhanss/guizhancraft/slimefun/items/machine/GuizhanMachineBlock.java
@@ -1,4 +1,0 @@
-package net.guizhanss.guizhancraft.slimefun.items.machine;
-
-public class GuizhanMachineBlock {
-}
