+++ src/main/resources/config.yml
@@ -1,3 +1,6 @@
+# DO NOT change this
+version: 2
+
 # Enable auto update from Guizhan Builds Page
 auto-update: true
 
@@ -7,5 +10,5 @@ auto-update: true
 ## since current Slimefun recipe system cannot recognize items with same id but different names
 lang: zh-CN
 
-# Enable research
-enable-research: true
+# Enable researches
+enable-researches: true
